Move paintPath to UserPathController and add constant for path width
