+++ src/main/java/controllers/user/UserMasterController.java
@@ -27,7 +27,6 @@ import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.text.Collator;
 import java.text.Normalizer;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
@@ -335,35 +334,6 @@ public abstract class UserMasterController
 		this.getScene().setRoot(userPath);
 	}
 
-	/**
-	 * Draw a simple path between the nodes in the given list
-	 *
-	 * @param directionNodes A list of the nodes in the path, in order
-	 */
-	public void paintPath(List<Node> directionNodes) {
-		this.directionsTextField.getChildren().clear();
-
-		//add kiosk to start of list
-		//directionNodes.add(0, this.kiosk);
-		if(directionNodes.size() <= 0) {
-			// TODO: Give an error message when no path is found
-			return;
-		}
-
-		// This can be any collection type;
-		Collection<Line> path = new HashSet<>();
-		for (int i=0; i < directionNodes.size()-1; ++i) {
-			Node here = directionNodes.get(i);
-			Node there = directionNodes.get(i + 1);
-			if (here.getFloor() == floor && here.getFloor() == there.getFloor()) {
-				Line line = new Line(here.getX(), here.getY(), there.getX(), there.getY());
-				path.add(line);
-			}
-		}
-		this.botPane.getChildren().setAll(path);
-
-
-	}
 
 	protected void changeFloor(int floor) {
 		this.switchFloors(floor);
